♻️🐎  only run explore cases pies query when rendered
