+++ src/packages/@ncigdc/containers/explore/ExplorePage.js
@@ -11,7 +11,6 @@ import MutationsTab from '@ncigdc/components/Explore/MutationsTab';
 import OncogridTab from '@ncigdc/components/Explore/OncogridTab';
 import CasesTab from '@ncigdc/components/Explore/CasesTab';
 import NoResultsMessage from '@ncigdc/components/NoResultsMessage';
-import ExploreCasesPies from '@ncigdc/components/TabPieCharts/ExploreCasesPies';
 import CaseAggregations from '@ncigdc/containers/explore/CaseAggregations';
 import GeneAggregations from '@ncigdc/containers/explore/GeneAggregations';
 import SSMAggregations from '@ncigdc/containers/explore/SSMAggregations';
@@ -119,7 +118,7 @@ export const ExplorePageComponent = (props: TProps) =>
             id: 'cases',
             text: `Cases (${props.viewer.explore.cases.hits.total.toLocaleString()})`,
             component: !!props.viewer.explore.cases.hits.total
-              ? <CasesTab pies={props.viewer.explore.cases.pies} />
+              ? <CasesTab />
               : <NoResultsMessage>No Cases Found.</NoResultsMessage>,
           },
           {
@@ -209,9 +208,6 @@ export const ExplorePageQuery = {
           cases {
             aggregations(filters: $filters aggregations_filter_themselves: false) {
               ${CaseAggregations.getFragment('aggregations')}
-            }
-            pies: aggregations(filters: $filters aggregations_filter_themselves: true) {
-              ${ExploreCasesPies.getFragment('aggregations')}
             }
             hits(first: $cases_size offset: $cases_offset filters: $filters score: $cases_score sort: $cases_sort) {
               total
